+++ src/rembus/settings.py
@@ -50,6 +50,7 @@ class Config:
         self.ack_timeout = cfg.get("ack_timeout", def_ack_timeout)
 
         self.db_attach = db_attach(name)
+        logger.info("db: %s", self.db_attach)
 
 
 def db_attach(router_id):
@@ -61,7 +62,6 @@ def db_attach(router_id):
         db_name = f"ducklake:{data_dir}.ducklake"
 
     if "DUCKDB_IGNORE_DATA_PATH" not in os.environ:
-        logger.debug("ATTACH '%s' AS rl (DATA_PATH '%s')", db_name, data_dir)
         return f"ATTACH '{db_name}' AS rl (DATA_PATH '{data_dir}')"
     else:
         return f"ATTACH '{db_name}' AS rl"